添加数据录入和性能预测后台

+++ src/main/java/com/kglab/lmcp/entity/graph/nodes/PropertyValueNode.java
@@ -22,7 +22,7 @@ public class PropertyValueNode implements NodeEntity {
     @GeneratedValue(UUIDStringGenerator.class)
     private final String uuid;
     /** 属性值 **/
-    private final double propValue;
+    private final float propValue;
 }
